requirements.txt
streamlit
python-dotenv
requests
pytest
openai